Rough first attempt at native windows support
I am starting with the first commit on this repo since it has way less in it and so I have less to port.
That has still led to me needing to polyfill lots of missing functionally missing from windows which I have done using a mix of cosmopolitan, gnulib(will be removed whenever possible) and a mmap implementation I found online that was specifically designed for mingw.
Only blink is working currently with blinkenlights totally removed, it will be readded however as this was again done to reduce the amount of stuff I had to work on.