Started adding actual tui
Many keyboard shortcuts are missing and breakpoints are disabled.
Additionally the output panel is not displayed due to a assert failing, the same thing happened on cygwin orignally so this should be easy to fix.
There is a bug when opening tui.exe on windows where DrainInput requires pressing quite a few chars before the tui appears, I am not currently sure why.