Fix freeze on tui start
It is no longer required to press tab to start the tui.
Setup log file on windows, outputs to wherever GetTempPathA says to in order of: $TMP, $TEMP $USERPROFILE, $WINDIR. Can be set via -L argument.
Enabled help screen activated with ? shortcut, has a visual bug in that the right hand side of the box has its char sprayed out to the left of the box.